internal/filetypes: speed up common case
Creating a `*build.File` value is somewhat expensive
in the general case, but the general case is fairly
unusual: it's much more common to have files
interpreted according to their extension only
with defaults for all other encoding attributes.

Make this operation more efficient (no CUE value
lookup required) for this common case by caching
the resulting `*build.File` values for each of the known
file extensions at init time.

While we're about it, change the existing "common case"
logic to use a check that matches the actual invariants of
the common case for CUE: specifically the default
encoding for CUE uses `form: ""` not `form: "schema"`: this
makes a difference in subsequent CLs in the checks
for whether a given parsed CUE file can be cached.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I46f8c823cca08a44a09dece8d5e9a3df55bcf928