internal/buildattr: stop at package clause and support parentheses
The existing logic was not behaving as expected: it did not stop when it
encountered a `package` clause.

Also, the expressions did not support parentheses, which seems like an
oversight: it's much less useful to have `&&` and `!` operators if it's
not possible to group subexpressions.

Fixes #3228.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ifed4419f4c89b6d52389c06aa960563e3f045a87
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1197530
Reviewed-by: Paul Jolly <paul@myitcv.io>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>