mod/modfile: use fixed language version for Fix
CUE currently uses the latest known language version when fixing legacy
module files.

This is probably incorrect, at least in the future, because the CUE
version used by such legacy modules is likely to be an older version and
so use older language features that might be deprecated or altered by a
newer language version.

To fix that, so we can always use a specific version (v0.9.0) when
fixing such files.

Fixes #3222.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ie525e069548c9b8714e72b058e2f8322cb94f5dd