mod/modload: ignore _test.cue files in dependencies
Currently files with a `_test.cue` suffix are recognized and ignored by
the `cue/load` code unless the Config.Tests field is true. However, as
there is no way to set that field from the command line, no way to
include a _test.cue file even when explicitly mentioned, and no
framework for writing tests in CUE, this is of very limited use.
However, it would be nice to provide proper test support in CUE in the
future, and the current behavior at least reserves a space in which that
can happen.

Thus, the dependency resolution code in mod/modload should match that of
cue/load, which ignores _test.cue files unless Config.Tests is true. It
seems like it would also be good to align with the current behavior of
_tool.cue files: they are ignored when outside the main module.

Also add a comment and make the logic a bit clearer
(shouldIncludePkgFile doesn't make a difference for packages in the main
module because they're added as part of the initial root set, ignoring
all build tags and including _test.cue and _tool.cue files.

Note: This differs from Go's treatment of test files: in Go, the
dependency resolution code includes enough dependencies to enable
running tests in any package directly or directly imported by the
non-test code, but does not include dependencies used for tests of
packages used only by external test code. For CUE, at least initially,
we will include only the dependencies needed by files in the main
module.

Fixes #3272.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ifd3bb59652ef3fa4739817f1f551a58a8b1f5996