mod/modload: ignore _test.cue files in dependencies
The dependency resolution code in mod/modload should match
that of cue/load, which ignores _test.cue files unless
Config.Tests is true.

Currently it always includes _test.cue files, but instead it
should include them only in the main module.

Also add a comment and make the logic a bit clearer
(shouldIncludePkgFile doesn't make a difference for packages
in the main module because they're added as part
of the initial root set, ignoring all build tags and including
_test.cue and _tool.cue files.

Fixes #3272.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ifd3bb59652ef3fa4739817f1f551a58a8b1f5996