cue: use Value.Err in examples using Compile APIs
The Context.CompileString docs say:

    The returned Value will represent an error, accessible through Err,
    if any error occurred.

Our two examples involving these APIs did not check for an error.
This can be OK if the user is sure an error won't happen,
but in general that's not a safe assumption to make.
Do the safer and recommended error check in the examples.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I1dd758d275f457db9eadc9aa6bf0faf7805fdb2b