cmd/cue: better error message for mismatched publish version
The error messages did not make it that clear what the issue was
when there's a version mismatch between the version argument
passed to `cue mod publish` and the major version in the
module file.

Example of old error message:

    cannot form module version: mismatched major version suffix in "mod.example/blah" (version v1.0.0)

Example of new message:

   publish version "v1.0.0" does not match implied major version "v0" in "cue.mod/module.cue"; must be v0.N.N

Also improve a couple of other error messages in this area, making clear that a canonical semver
is required.

Fixes #3225

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Id84166be9e5c97216d170a8968ec44354a56cb82