cue/literal: use strconv.IsGraphic
This code was a direct copy of strconv.IsGraphic with no changes,
so there isn't a need to duplicate it.
We already used strconv.IsPrint in this Go file,
and we already used both APIs in other Go packages in our module.

While here, as suggested by Matthew, use parentheses to make the
boolean logic clear with respect to operator precedence.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I4d4f6416e8a265fee89b3ed357fdb3d4771d0e7f