cmd/cue: add tests for module path errors
This adds some tests for some errors that will be improved in
the next CL in this chain.

We also add tests for some of the module path checking functions
that were not properly unit tested previously so we can be sure
that their behavior does not change (an earlier iteration _did_
unexpectedly change the behavior of `CheckImportPath`).

As the test table is shared between several test functions (useful,
as it's nice to see the results for each check function across all
the input data), we needed to update the `tdtest` package to support
tables defined at the top level. We include this change in this CL
because then it's clear that it actually works.

For #3022

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I7cd5da04858c7610dcadfa6a28dff8a0f2e6ea77