pkg/struct: count required fields in MinFields and MaxFields
While a required field is technically not a regular field that counts
as being present, a required field can only be completed once it is
declared as a regular field. For example, it makes little sense for

    {a!: 1, b!: 2}

to satisfy the constraint

    struct.MaxFields(1)

because no complete configuration does ever satisfy the constraint.
For the sake of consistency and to match a user's intuition,
count required fields as if they were regular fields.

This was an oversight from when required fields were added.
The pkg/structs code was correct when we just had optional fields,
but now that we have required fields too, its logic needs tweaking.

Note that the IsOptional method can go as well now.
The extra loop to find the same label again is unnecessary,
and we do want to match on ArcRequired now.

Fixes #3141.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I98c7d39be43a140e5eb20aa2b96943f2e43fd562