internal/mod/modimports: remove duplicate test txtar file entry
Switching from our own txtarfs package to upstream's new txtar.FS API
surfaced the fact that TestPackageFiles had duplicate file entries
for y.cue, leading to an error as the new API is stricter.

Our existing logic uses the first entry when there are duplicates,
so delete the second. This is verified by the test failing
when the first entry is deleted instead.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Ie37a9200b5ed854379ba77884779fbf951cb35bf