all: replace internal/txtarfs with txtar.FS
https://go.dev/issue/44158 was resolved in x/tools master yesterday,
and the new API behaves in a very similar way to our previous one.

One major difference is that the API errors upfront if any txtar paths
are not valid filesystem names, whereas our previous implementation
only errored when invalid paths were opened.
For this reason, the new API returns an error where ours didn't,
so a minor refactor is needed across the test files.

This pushes us to depending on x/tools master as of yesterday,
but it seems safe given that the API has been approved for months
and is very unlikely to change before the next x/tools release.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I6e1974c4a77be4207830d78f579d21cf445369a6