drop go1.21, start testing on go1.23rc2
In preparation for CUE v0.10.0, which will be released around the same
time that Go 1.23.0 is released in mid-August.
This lets us get better early testing for Go 1.23, as well as start
taking advantage of some of the features in Go 1.22.

While here, move the Go version matrix declaration to repo.cue,
as we declared versions of Go which were only used for the matrix list.
It is more intuitive to declare the list directly, which also lets us
add or remove versions more easily, such as RCs for early testing.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Ia66d9d4d0428681cee7352301b0bea41d7aa0dfe
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1198206
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Paul Jolly <paul@myitcv.io>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>