cmd/cue,internal/buildattr: add tests for `@ignore` tags
This CL adds some tests for the up-coming `@ignore` functionality so
that we can see the difference when it is implemented.

Note that the syntax is just the general CUE attribute syntax so
it's valid, just ignored by the loader as yet.

For #2962.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I54e4b098bd425d80d4dfa18b1086a0181cc2dee7
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1198248
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>