internal/buildattr: implement `@ignore` attributes
This adds support for an `@ignore()` attribute
to enable a file to be unconditionally ignored.

We decided against following Go's lead here,
which would have been `@if(ignore)` because there is no
existing convention like Go had, and because this is
sufficiently different from regular build tags that it seems like it
justifies another attribute. Also because `@ignore()` seems to
read a bit better than `@if(ignore)`.

Fixes #2962.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I7ee84cea2ffce6765bd7667feb78c4d44e0254ea