cue/cmd: ensure flags are added if they are used
Problem: trim used the dry-run flag but did not add it.

This change ensures that when a value is fetched for any flag, we ensure
that the flag has been added to the flagset. We panic if it has not
been.

Although better than the status quo, this is not quite perfect because
it is only when the value is fetched at runtime that we perform this
checking, and not all code paths require every flag value. So in the
case of trim, if trim exits early due to some bug or divergence, then
the dry-run flag value is never inspected and so we never do this
checking.

Change-Id: I21263c4da9f7550e994d4bd4cbe0195f8946b2c3
Signed-off-by: Matthew Sackman <matthew@cue.works>