internal/ci: ensure commit messages do not @-mention users
Parse the commit message as GitHub-flavored markdown to only search
for @-mentions which would be linkified by GitHub, that is,
not within backticks or block quotes.

Adding goldmark as a Go dependency is okay given that it has zero
dependencies itself, and that Go's lazy module loading means
that any Go users of CUE as a library won't download it.

Add positive and negative test cases as well.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Ie2c86998bc11384a6d2e6c96c40027a519e233df
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1198294
Reviewed-by: Paul Jolly <paul@myitcv.io>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>