internal/_e2e: remove unused import
Somehow in https://cuelang.org/cl/1198292 I added an unused import
to script_test.go, and it wasn't picked up by the trybots
as the e2e tests are only run on commits pushed to master.

A follow-up commit will teach the trybots to check that internal/_e2e
can be compiled as a test package, as this is not the first,
nor will it be the last, that we mess up its build by accident
and forget that CI on Gerrit does not notice.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I557ac8c5e3ed04480322bf9a7ea6daf1a3884d1d
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1198297
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Paul Jolly <paul@myitcv.io>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>