cue: add a regression test for a default elimination bug fixed in evalv3
The old evaluator does not correctly follow the rewrite rules for
default values:

    https://cuelang.org/docs/reference/spec/#default-values

It then makes a further mistake by output different values for JSON or
Yaml.

The new evaluator correctly determines that the field is incomplete
because it is left with 3 | >=3.

Add a regression test to lock in the behaviour of the new evaluator.

Closes #2916.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I9a5151df9702dfa46271edf543978bce6b1ec3a1