internal/ci: ensure internal/_e2e isn't broken in CI
I have broken the build of the internal/_e2e tests more than once,
most recently fixed in https://cuelang.org/cl/1198297,
as we only actually run these tests on pushes to protected branches.

However, it's easy to build the tests in a CI run before merging,
and that will catch most of these mistakes which can happen as a result
of refactors or other API changes which internal/_e2e uses.

DO NOT MERGE: breaking _e2e on purpose to test run

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I529ed3b1d52e86ef1f9c41744ebc2d97ef3d77a2