all: start using Go 1.22's loop var scoping and reflect.TypeFor
The `x := x` trick to make a copy of the shared for loop variable
is no longer necessary as of Go 1.22 and later,
as explained in https://go.dev/doc/go1.22#language.

reflect.TypeFor is a nice generic wrapper for reflect.TypeOf
where the type in question is known at compile time.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I76450d91efc8f52d89211b1cbb97838730da15f4