cmd/cue: point "help get" text at CUE-specific cmd
This updates the "cue help get" text to make it clear that "cue get"
doesn't interact with CUE dependencies. For CUE modules, it updates two
help texts by pointing the reader onwards to "cue help modules", and
from there to "cue help mod".

Also: fix a typo (which was the initial motivation for this edit)

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>
Change-Id: I2a980be042fe544c95a71c8b46a20616cb7c4bd6
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1198447
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>