lsp: expose LineCol8Location on integration.Env
This convenience mechanism allows us to specify ranges or end-less
positions in a file in a human-readable shorthand form:

    file:5:2-4:3      // full range
    file:5:2          // end pos is implied from start pos

In CUE we need to match against multiple locations, and so doing
regexep-based searches over test files is error prone. Best to be
specific on the multiple locations with actual locations, in a format
that is well understood by the reader.

It's also possible to use editor based tooling for calculating the range
string based on a visual selection, relative to the current txtar file
header. That is left as an exercise for the reader.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I0dafeac7906a131510b10d7a123171d6357318e1