pkg: add test cases for ** in file patterns or globs
Right now we follow the Go semantics, where it is interpreted as two
`*` symbols next to each other, thus behaving like a single star symbol
as the second one will never match anything.

The following commit will change this behavior, to be shown in the
changes made to these added test cases.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I521e896c7349f0320bbe5afc63c45b78a7d8a9eb