pkg: forbid `**` in path.Match and tool/file.Glob
Currently the double-star pattern term `**` behaves like a star `*`,
as one star matches any sequence of non-separator characters
and the second star matches nothing.

This is confusing, as one may write the pattern `**/*.json` and it can
match strings like `foo/bar.json`, making it seem like `**` works as
a double star that would also match `foo.json` or `foo/bar/baz.json`
when in fact it does not.

If we choose to support `**` patterns in the future to match separators,
this would mean changing the behavior for existing users in a subtle way
which could lead to unintended behavior or latent bugs.
It is clearer for existing users to treat `**` as an invalid pattern,
which would allow us to add the feature in the future without concern
about existing users already using such patterns without error.

For #1919.
Fixes #3315.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Ifeb81f4818a863cd0e4c5c13bc0fc2ac54a0f22f