cue/interpreter/embed: forbid `**` via pkg/path.Match
Our pkg/path.Match func mimics Go's own path.Match in that it can be
used as a validator via `Match(pattern, "")`. This is better than a
`strings.Contains(pattern, "**")` not only because we get consistent
error messages, but also because Contains gives false positives
on valid patterns such as `\**`, which escapes the first star,
or `[x**y]`, where the stars are just part of a character set.

Fix the code, and add more tests that demonstrate that embedding
does support a glob with `\**` now.

This brings embed in line with pkg/path.Match and tool/file.Glob
in terms of what pattern matching syntax and features it supports.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I0ed6fcab9426c6392fb18491a3f656f29c56b41e
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1198692
Reviewed-by: Roger Peppe <rogpeppe@gmail.com>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>