encoding/jsonschema: port txtar tests to cuetxtar
This saves about 60 lines of code to deal with loading the txtar files
as well as updating output files following CUE_UPDATE and so on,
but most importantly, it unlocks the ability to test with evalv3 too.

For #3351.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I3f9a368391c4510d2f5a389da3ac7babb376e036
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1198734
Reviewed-by: Roger Peppe <rogpeppe@gmail.com>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>