encoding/toml: support decoding keys named "_"
We can't blindly create an ast.Ident with such a name, as that then
refers to the "top" identifier in CUE, which cannot be used as a label:

    foo: cannot use _ as label:

We hadn't noticed this with other keys which need quoting such as "a-b"
because sticking the unquoted string in ast.Ident.Name caused cue/format
to quote the identifier correctly. It does not do this for "_" as it is
a valid identifier - it just represents top, and not a literal string.

Fixes #3350.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: If0f2c2fd6c0d984a7c6c30e4667ab8ed097aa17b