cmd/cue: add TOML to a couple of help topics I forgot about
Neither `cue help filetypes` nor `cue help vet` listed it next to YAML,
which was an omission on my part when I added TOML support to cmd/cue.

This made me wonder for a second whether `cue vet` would support
TOML files out of the box already, so I added a test to confirm this.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I3f0d06f3234c5b4d566c47a912d062b5b16c3a1e