internal/core/adt: fix panic triggered by using wrong condition
Use the isInitialized() instead of the corresponding flag in
nodeContext:  a finalized node, may currently still have an
uninitialized nodeContext.

This bug was exposed by a code path that is itself exposing a
bug. We add tests for this, but also add a test case that is
not erroneous, but exposes this code path so that this fix
will remain tested even if the triggering bug is fixed.

Fixes #3330

Signed-off-by: Marcel van Lohuizen <mpvl@gmail.com>
Change-Id: Ibe7d130bf66a22b14df7dd4012fd346649280419
Reviewed-on: https://cue.gerrithub.io/c/cue-lang/cue/+/1198860
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>