internal/encoding/yaml: decode empty inputs as "null"
The YAML 1.2.2 spec says:

    YAML allows the node content to be omitted in many cases. Nodes with
    empty content are interpreted as if they were plain scalars with an
    empty value. Such nodes are commonly resolved to a “null” value.

And indeed, multiple of their examples show empty inputs decoding
as null, such as

    ---
    ---
    # Empty
    ...

resulting in two "null" values.

The old YAML decoder seems to have wanted this behavior for years,
but as my TODO on the new decoder explains, it returned a "null" literal
alongside an io.EOF error, and the consumer discarded the literal.

This resolves the mishandling of empty inputs, which reached io.EOF
directly without providing any decoded CUE expression,
and also resolves all panics that this caused in cmd/cue.

Fixes #1790.
Fixes #2714.
Fixes #3337.
Closes #1807.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Iaa3356c7eee5e84dc51386a057a470138e1ac3ce