update golang.org/x/... and ociregistry for rc.1
To bring in the latest fixes from the x repos, particularly as Go 1.23
is around the corner, and to bring in the init time cost reductions
we just merged into the oci repository.

Note that the cue repo does not use any of the ociregistry IsValid APIs
which have been deprecated in the latest ociregistry patch.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I665e6e47f068b08100c9e1e743ed147295ddc7b4