encoding/toml: mimic indented CUE multi-line strings in test strings
This way, when we start reporting positions in errors,
we use line and column numbers that visually make sense for the reader,
rather than always beginning from the second line and indentation.

Note that we indent the closing quote by one more level,
to match CUE's indentation semantics for multiline strings.

No change in the test results here; this change is separated to make
reviewing the following patches easier.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I1bf6f9422b623c4c6458290b448efde9330b7d88