encoding/toml: support error positions
Including filenames as well as positions from the toml parser
as well as our own decoding logic.

We use the cue/errors API for the sake of returning error positions
as part of error values, and not just as part of an opaque error string.
Note that this requires the test code to use errors.Details to render
error positions, as the Error method does not by default.

Fixes #3339.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I9f035bad95fd0bf8990033a91fd0393c544a0f79