cmd/cue: add a test for embedding symbolic links
Valid symbolic links are currently allowed, but that will change
in the next commit. This test allows us to see the change in behavior.

While here, add more error test cases for embedding as well.
In particular, the "cannot embed directories" and invalid filetype
error code paths were not covered in the tests.

For #3299.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I5f80b1dc81e2a2bd13bd084be74cabec6f205ba4