cue/interpreter/embed: forbid embedding symlinks
At least for now, we aren't positive that we want to support them,
so it is easier to release the `@embed` experiment in v0.10.0
without support for symlinks, and consider adding it at a later time
as users request the feature. On the flip side, removing a feature
later on would be a breaking change that would be harder to do safely.

Moreover, note how Go has never supported embedding irregular files,
and it seems like that hasn't been a problem so far.

Unfortunately we cannot accomplish this with io/fs.FS just yet,
as io/fs.SymlinkFS has not been implemented and released still.
However, we always use os.DirFS, so we can sidestep the io/fs.FS
and reach straight for os.Lstat. Leave TODOs about cleaning this up.

Fixes #3299.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Icc13f21d421a4241d59246832de059a51e795d90