cue/interpreter/embed: forbid embedding symlinks
At least for now, we aren't positive that we want to support them,
so it is easier to release the `@embed` experiment in v0.10.0
without support for symlinks, and consider adding it at a later time
as users request the feature. On the flip side, removing a feature
later on would be a breaking change that would be harder to do safely.

Moreover, note how Go has never supported embedding irregular files,
and it seems like that hasn't been a problem so far.

Unfortunately we cannot accomplish this with io/fs.FS just yet,
as io/fs.SymlinkFS has not been implemented and released still.
However, we always use os.DirFS, so we can sidestep the io/fs.FS
and reach straight for os.Lstat. Leave TODOs about cleaning this up.

Note that we only add a test case with a broken symlink here,
as that used to result in OS-specific error messages before
we started forbidding symbolic links:

    @embed: failed to stat symlink-broken.json: stat symlink-broken.json: no such file or directory
    @embed: failed to stat symlink-broken.json: CreateFile symlink-broken.json: The system cannot find the file specified.

While here, make the "cannot embed" errors mention the filename.

Fixes #3299.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Icc13f21d421a4241d59246832de059a51e795d90