internal/core/adt: show # only on original defintion node
isDef is marked as # in the graphical debugger. As isDef
is propagated up, all ancestor nodes are propagated.
For debugging purpososes, this makes it hard to see where
the isDef originated, which is generally more in line what
we are interested in.

This change now only shows # on nodes where it was
originally set.

Signed-off-by: Marcel van Lohuizen <mpvl@gmail.com>
Change-Id: Iae2cbf287621d4b70b80fe3595ee43a42d236860
Reviewed-on: https://cue.gerrithub.io/c/cue-lang/cue/+/1199269
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>