internal/core/adt: split conjunct trees based on prefix
Before this change, any conjunct referred to was inserted
at the insertion point in full. Top-level conjuncts have
a tree representing the origin of their definition and
embedding. This caused the closedness rules to be repeated
when they should not, resulting in spurious "field not
allowed" errors.

The new algorithm splits a conjunct into branches with
roots that are shared with the insertion point removed.
This ensures that references that are in scope of a
definition do not get spuriously inserted.

Fixes #3330
Fixes #3331

Signed-off-by: Marcel van Lohuizen <mpvl@gmail.com>
Change-Id: I7a573f9576d3c46374647045d8ec226a1cd543b3