encoding/jsonschema: fix decoding of `$id`
When decoding "$id", jsonschema decoder mistakenly assumed that
the schema is an object and added an object constraint with the id tag.

This change postpones the attachment of id tag until an object
constraint is added or the schema is finalizing.

Fixes #3361

Change-Id: I7829af165af726524d48022fc87ed538b7e451cd
Signed-off-by: haoqixu <hq.xu0o0@gmail.com>
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue/+/1199139
Unity-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Roger Peppe <rogpeppe@gmail.com>
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>